Fix for gfc-139
Implemented an admin site view that will display the most participations
in any user activity for any given month. In the 'change list' view for
the 'user activity' model, there is a new button in the actions list,
namely 'Activity Stats'. When clicking this button, you get a view where
you can select the activity, and for which month you want to view it.
Results will be displayed below the criteria.